sql로그추가및닉네임중복체크추가

+++ src/main/java/com/java/dogle/service/member/MemberService.java
@@ -3,5 +3,5 @@ package com.java.dogle.service.member;
 import com.java.dogle.vo.member.MemberVO;
 
 public interface MemberService {
-	public long chkNicknameDup(MemberVO param);
+	public int chkNicknameDup(MemberVO param);
 }
